Refactored util package design.
