+++ core/src/main/java/be/kuleuven/cs/gridflex/util/ArrayUtils.java
@@ -1,4 +1,4 @@
-package be.kuleuven.cs.gridflex.domain.util;
+package be.kuleuven.cs.gridflex.util;
 
 /**
  * Utility class for array calculations.
