+++ core/src/main/java/be/kuleuven/cs/gridflex/util/MathUtils.java
@@ -1,4 +1,4 @@
-package be.kuleuven.cs.gridflex.domain.util;
+package be.kuleuven.cs.gridflex.util;
 
 /**
  * Class for math-based utility functions.
