+++ core/src/main/java/be/kuleuven/cs/gridflex/util/listener/MultiplexListener.java
@@ -1,4 +1,4 @@
-package be.kuleuven.cs.gridflex.domain.util.listener;
+package be.kuleuven.cs.gridflex.util.listener;
 
 /**
  * Listener that multiplexes other listeners.
