+++ core/src/main/java/be/kuleuven/cs/gridflex/util/listener/package-info.java
@@ -1,9 +1,9 @@
 /**
- * Package for entities concerning visitors.
+ * Pakcage for implementing observer pattern-ish construct with listeners.
  * 
  * @author Kristof Coninx (kristof.coninx AT cs.kuleuven.be)
  */
 @NonNullByDefault
-package be.kuleuven.cs.gridflex.domain.util.visitor;
+package be.kuleuven.cs.gridflex.util.listener;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
