+++ core/src/main/java/be/kuleuven/cs/gridflex/util/visitor/package-info.java
@@ -1,9 +1,9 @@
 /**
- * Pakcage for implementing observer pattern-ish construct with listeners.
+ * Package for entities concerning visitors.
  * 
  * @author Kristof Coninx (kristof.coninx AT cs.kuleuven.be)
  */
 @NonNullByDefault
-package be.kuleuven.cs.gridflex.domain.util.listener;
+package be.kuleuven.cs.gridflex.util.visitor;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
