+++ core/src/test/java/be/kuleuven/cs/gridflex/util/CollectionUtilsTest.java
@@ -1,4 +1,4 @@
-package be.kuleuven.cs.gridflex.domain.util;
+package be.kuleuven.cs.gridflex.util;
 
 import com.google.common.collect.Lists;
 import org.junit.Before;
@@ -11,14 +11,14 @@ import static org.junit.Assert.assertEquals;
 
 public class CollectionUtilsTest {
 
-    List<FlexBid> bids = Lists.newArrayList();
+    List<Bid> bids = Lists.newArrayList();
 
     @Before
     public void setUp() throws Exception {
-        bids.add(new FlexBid(FlexTuple.NONE, 1));
-        bids.add(new FlexBid(FlexTuple.NONE, 2));
-        bids.add(new FlexBid(FlexTuple.NONE, 4));
-        bids.add(new FlexBid(FlexTuple.NONE, 3));
+        bids.add(new Bid(1));
+        bids.add(new Bid(2));
+        bids.add(new Bid(4));
+        bids.add(new Bid(3));
     }
 
     @Test
@@ -29,13 +29,13 @@ public class CollectionUtilsTest {
 
     @Test
     public void testMaxEmpty() {
-        assertEquals(0, CollectionUtils.max(new ArrayList<FlexBid>(),
+        assertEquals(0, CollectionUtils.max(new ArrayList<Bid>(),
                 input -> input.getValuation()), 0);
     }
 
     @Test
     public void testArgMax() {
-        bids.add(new FlexBid(FlexTuple.NONE, 7));
+        bids.add(new Bid(7));
         assertEquals(bids.get(4),
                 CollectionUtils.argMax(bids, input -> input.getValuation()));
     }
@@ -49,11 +49,11 @@ public class CollectionUtilsTest {
     @Test
     public void testCorner() {
         bids = Lists.newArrayList();
-        bids.add(new FlexBid(FlexTuple.NONE, 10));
-        bids.add(new FlexBid(FlexTuple.NONE, 2));
-        bids.add(new FlexBid(FlexTuple.NONE, 4));
-        bids.add(new FlexBid(FlexTuple.NONE, 3));
-        bids.add(new FlexBid(FlexTuple.NONE, 7));
+        bids.add(new Bid(10));
+        bids.add(new Bid(2));
+        bids.add(new Bid(4));
+        bids.add(new Bid(3));
+        bids.add(new Bid(7));
 
         assertEquals(bids.get(0).getValuation(),
                 CollectionUtils.max(bids, input -> input.getValuation()));
@@ -63,5 +63,17 @@ public class CollectionUtilsTest {
 
         assertEquals(26,
                 CollectionUtils.sum(bids, input -> input.getValuation()));
+    }
+
+    class Bid {
+        private final int valuation;
+
+        Bid(int bid) {
+            this.valuation = bid;
+        }
+
+        public int getValuation() {
+            return valuation;
+        }
     }
 }
