+++ domain/pom.xml
@@ -20,10 +20,6 @@
             <version>0.1.4-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>com.googlecode.combinatoricslib</groupId>
-            <artifactId>combinatoricslib</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.google.auto.value</groupId>
             <artifactId>auto-value</artifactId>
             <scope>provided</scope>
