+++ domain/src/main/java/be/kuleuven/cs/gridflex/domain/energy/tso/simple/CopperplateTSO.java
@@ -3,9 +3,9 @@ package be.kuleuven.cs.gridflex.domain.energy.tso.simple;
 import be.kuleuven.cs.gridflex.domain.energy.consumption.EnergyConsumptionTrackable;
 import be.kuleuven.cs.gridflex.domain.energy.generation.EnergyProductionTrackable;
 import be.kuleuven.cs.gridflex.domain.energy.tso.BalancingSignal;
-import be.kuleuven.cs.gridflex.domain.util.listener.Listener;
-import be.kuleuven.cs.gridflex.domain.util.listener.MultiplexListener;
-import be.kuleuven.cs.gridflex.domain.util.listener.NoopListener;
+import be.kuleuven.cs.gridflex.util.listener.Listener;
+import be.kuleuven.cs.gridflex.util.listener.MultiplexListener;
+import be.kuleuven.cs.gridflex.util.listener.NoopListener;
 import be.kuleuven.cs.gridflex.simulation.SimulationComponent;
 import be.kuleuven.cs.gridflex.simulation.SimulationContext;
 import com.google.common.collect.Lists;
