+++ domain/src/main/java/be/kuleuven/cs/gridflex/domain/site/SiteImpl.java
@@ -3,9 +3,9 @@ package be.kuleuven.cs.gridflex.domain.site;
 import be.kuleuven.cs.gridflex.domain.process.FlexProcess;
 import be.kuleuven.cs.gridflex.domain.resource.Resource;
 import be.kuleuven.cs.gridflex.domain.util.FlexTuple;
-import be.kuleuven.cs.gridflex.domain.util.listener.Listener;
-import be.kuleuven.cs.gridflex.domain.util.listener.MultiplexListener;
-import be.kuleuven.cs.gridflex.domain.util.listener.NoopListener;
+import be.kuleuven.cs.gridflex.util.listener.Listener;
+import be.kuleuven.cs.gridflex.util.listener.MultiplexListener;
+import be.kuleuven.cs.gridflex.util.listener.NoopListener;
 import be.kuleuven.cs.gridflex.simulation.SimulationComponent;
 import be.kuleuven.cs.gridflex.simulation.SimulationContext;
 import com.google.common.collect.LinkedListMultimap;
