+++ domain/src/main/java/be/kuleuven/cs/gridflex/domain/site/SiteSimulation.java
@@ -4,9 +4,9 @@ import be.kuleuven.cs.gridflex.domain.process.FlexProcess;
 import be.kuleuven.cs.gridflex.domain.resource.Resource;
 import be.kuleuven.cs.gridflex.domain.resource.ResourceFactory;
 import be.kuleuven.cs.gridflex.domain.util.FlexTuple;
-import be.kuleuven.cs.gridflex.domain.util.listener.Listener;
-import be.kuleuven.cs.gridflex.domain.util.listener.MultiplexListener;
-import be.kuleuven.cs.gridflex.domain.util.listener.NoopListener;
+import be.kuleuven.cs.gridflex.util.listener.Listener;
+import be.kuleuven.cs.gridflex.util.listener.MultiplexListener;
+import be.kuleuven.cs.gridflex.util.listener.NoopListener;
 import be.kuleuven.cs.gridflex.simulation.SimulationContext;
 import be.kuleuven.cs.gridflex.simulation.UIDGenerator;
 import com.google.common.collect.Lists;
