+++ domain/src/test/java/DomainTests.java
@@ -21,13 +21,9 @@ import be.kuleuven.cs.gridflex.domain.site.EquidistantSiteSimulationTest;
 import be.kuleuven.cs.gridflex.domain.site.SiteSimulationTest;
 import be.kuleuven.cs.gridflex.domain.site.SiteTest;
 import be.kuleuven.cs.gridflex.domain.util.BufferTest;
-import be.kuleuven.cs.gridflex.domain.util.CollectionUtilsTest;
-import be.kuleuven.cs.gridflex.domain.util.NPermuteAndCombinerTest;
-import be.kuleuven.cs.gridflex.domain.util.TrapzPosTest;
 import be.kuleuven.cs.gridflex.domain.util.data.FlexTupleTest;
 import be.kuleuven.cs.gridflex.domain.util.data.WindSpeedForecastMultiHorizonErrorDistributionTest;
 import be.kuleuven.cs.gridflex.domain.util.data.profiles.CongestionProfileTest;
-import be.kuleuven.cs.gridflex.domain.util.listener.ListenerTest;
 import be.kuleuven.cs.gridflex.domain.workstation.WorkstationTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
@@ -39,12 +35,12 @@ import org.junit.runners.Suite.SuiteClasses;
         FlexTupleTest.class, SiteTest.class, AggregatorTest.class,
         BRPAggregatorTest.class, IndependentAggregatorTest.class,
         ReactiveMechanismAggregatorTest.class, RandomTSOTest.class,
-        NPermuteAndCombinerTest.class, CopperPlateTSOTest.class,
+        CopperPlateTSOTest.class,
         OutputGeneratorTest.class, SiteSimulationTest.class,
         BalancingAuctionTSOTest.class, BalancingTSOTest.class,
-        EquidistantSiteSimulationTest.class, ListenerTest.class,
-        NominationTest.class, CongestionProfileTest.class, TrapzPosTest.class,
-        CollectionUtilsTest.class, DSMProposalTest.class, DSMPartnerTest.class,
+        EquidistantSiteSimulationTest.class,
+        NominationTest.class, CongestionProfileTest.class,
+        DSMProposalTest.class, DSMPartnerTest.class,
         DSOIntegrationTest.class, TurbineSpecificationTest.class,
         TurbineConvertorTest.class, WindSpeedForecastMultiHorizonErrorDistributionTest.class,
         BudgetTrackerTest.class })
